+++ src/commands/SyncLocalWithBucketCommand.ts
@@ -17,6 +17,7 @@ export type SyncLocalWithBucketCommandInput = {
   localDir: string;
   dryRun?: boolean;
   del?: boolean;
+  deleteExcluded?: boolean;
   sizeOnly?: boolean;
   relocations?: Relocation[];
   filters?: Filter[];
@@ -37,6 +38,7 @@ export class SyncLocalWithBucketCommand {
   localDir: string;
   dryRun: boolean;
   del: boolean;
+  deleteExcluded: boolean;
   sizeOnly: boolean;
   relocations: Relocation[];
   filters: Filter[];
@@ -50,6 +52,7 @@ export class SyncLocalWithBucketCommand {
     this.localDir = input.localDir;
     this.dryRun = input.dryRun ?? false;
     this.del = input.del ?? false;
+    this.deleteExcluded = input.deleteExcluded ?? false;
     this.sizeOnly = input.sizeOnly ?? false;
     this.relocations = input.relocations ?? [];
     this.filters = input.filters ?? [];
@@ -75,20 +78,14 @@ export class SyncLocalWithBucketCommand {
             : currentPath,
         ...this.relocations,
       ];
-    sourceObjects.forEach((sourceObject) =>
-      sourceObject.applyFilters(this.filters)
-    );
-    const includedSourceObjects = sourceObjects.filter(
-      (sourceObject) => sourceObject.isIncluded
-    );
-    includedSourceObjects.forEach((sourceObject) => {
+    sourceObjects.forEach((sourceObject) => {
+      sourceObject.applyFilters(this.filters);
       sourceObject.applyRelocations(this.relocations);
     });
-    const diff = SyncObject.diff(
-      includedSourceObjects,
-      targetObjects,
-      this.sizeOnly
-    );
+    const diff = SyncObject.diff(sourceObjects, targetObjects, {
+      sizeOnly: this.sizeOnly,
+      deleteExcluded: this.deleteExcluded,
+    });
     const commands = [];
     if (!this.dryRun) {
       commands.push(